refactored types and validators (more), fixed model verification
